relocated, page components
