+++ src/pages/CreateUser.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import UserForm from '../components/UserForm'
+import UserForm from '../components/users/UserForm'
 import { Typography, makeStyles } from '@material-ui/core'
 
 let useStyles = makeStyles((theme) => {
